+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add isDynamicLibraryChord flag to chord model
 - support WebUSB fallback
 
 ## [0.0.50](https://github.com/andy23512/tangent-cc-lib/releases/tag/0.0.50)